add chane min and max value function
